requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-autoslug-iplweb==1.9.4.dev0
django-braces==1.14.0
django-crispy-forms==1.9.1
django-hitcount==1.3.2
django-model-utils==4.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-star-ratings==0.9.1
django-taggit==1.3.0
djangorestframework==3.11.0
idna==2.9
mysqlclient==1.4.6
oauthlib==3.1.0
persiantools==1.4.1
Pillow==7.1.2
progressbar2==3.51.4
python-decouple==3.3
python-utils==2.4.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
Unidecode==1.1.1
urllib3==1.25.9
